pyproject.toml
[project]
name = "ai-class-400-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.0.0",
    "pytest>=9.0.2",
    "httpx>=0.28.1",
]
